requirements.txt
appdirs==1.4.3
attrs==19.3.0
black==19.10b0
Click==7.0
numpy==1.18.1
pandas==1.0.1
pathspec==0.7.0
psycopg2==2.8.4
python-dateutil==2.8.1
python-dotenv==0.11.0
pytz==2019.3
regex==2020.2.20
six==1.14.0
SQLAlchemy==1.3.13
toml==0.10.0
typed-ast==1.4.1
